Fixed custom mesh addition to be properly undoable
Custom mesh needs to specify a default mesh when first created, instead
of popping up a dialog immediately to select source.
Primary reason for this change is that undo/redo logic is much simpler,
when adding a mesh is actually a single event (add custom mesh entity)
rather than two different kinds of events (add and change source url),
that would have to be conflated into a single undo command.

Also changed the EditorUtils class to be purely static class, both so
that it is more accessible from various parts of the editor code and
enforce the idea that it is a collection of standalone utility
functions.

Change-Id: If37a2420269759b170dbe9f5930ec731237422f9
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@theqtcompany.com>