Fix item remove, insert, and duplication
Remove now duplicates the old entity and stores the duplicate,
since Qt3D has trouble letting go of the control over the entities
it once had. The removed entity is always deleted.

Changed InsertEntityCommand to store entity name instead of the
entity pointer, as the entity can get destroyed now.

Refactored item duplication code to support the remove, which includes
not duplicating internal components, as those will get recreated
when the duplicate is added to scene. Also removed name generation for
duplicated components, as we are not interested in those. Only entity
names are relevant.

Change-Id: I06cf8660722c74f7a16f3f8b31bb187f08ac66eb
Reviewed-by: Tomi Korpipää <tomi.korpipaa@theqtcompany.com>