Use QApplication again with standalone application
Introducing creator plugin caused standalone app menu style to change,
because QGuiApplication was used instead of QApplication.
The menu icons were also not working with the new style.
When launching via creator, the menus are on the old style, as
creator is based on QApplication, so revert the standalone to same
style.

Change-Id: I8b169f04e38dcfc1f16b39730aa033e96b38b87b
Reviewed-by: Tomi Korpipää <tomi.korpipaa@theqtcompany.com>